% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afinn_111.R
\docType{data}
\name{afinn_111}
\alias{afinn_111}
\title{COVID-19 pandemic related keywords}
\format{
A list (dataframe)
}
\usage{
afinn_111
}
\description{
A lst of lexicon of English words rated for valence with
an integer between minus five (negative) and plus five (positive).
The words have been manually labeled by Finn Årup Nielsen in 2009-2011.
}
\keyword{datasets}
