% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_keys}
\alias{covid_keys}
\title{keywords relating to COVID-19 pandemics}
\format{
A dataframe containing one variable:
\itemize{
\item keys: list of keywords
}
}
\usage{
covid_keys
}
\description{
A list of keywords relating to the COVID-19 pandemic
}
\keyword{datasets}
