% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{debate_dtd}
\alias{debate_dtd}
\title{Comments on a video of a political debate.}
\format{
A dataframe containing one variable
\itemize{
\item text: individual text records
}
}
\usage{
debate_dtd
}
\description{
A DTD containing individual comments on a video
showing the first debate between two US presidential
nominees (Donald Trump and Hillary Clinton)
in Sept. 2016. (Credit: NBC News).
}
\details{
The DTD only include the comments within the first 24hrs
in which the video was posted. All individual comments
in which the names of both candidates are mentioned
are filtered out.
}
\keyword{datasets}
