% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{facility_keys}
\alias{facility_keys}
\title{Keywords relating to facilities at train stations}
\format{
A dataframe containing one variable:
\itemize{
\item keys: list of keywords
}
}
\usage{
facility_keys
}
\description{
List of facilities at the Piccadilly Train
Station (Manchester)
}
\keyword{datasets}
