% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi_impact.R
\name{opi_impact}
\alias{opi_impact}
\title{To assess the impact of subject B on subject A}
\usage{
opi_impact(textdoc, sec_keywords=NULL, metric = 1,
fun = NULL, nsim = 99, alternative="two.sided", pplot=FALSE)
}
\arguments{
\item{textdoc}{A collection (dataframe) of individual text
records, such as tweets or Facebook posts. The first column
of the dataframe must contain the text records.}

\item{sec_keywords}{\link{list} A one column  dataframe (of any
number of rows) containing a list of keywords relating to
the secondary subject (\code{B}).}

\item{metric}{\link{integer} Metric to utilize for the calculation
of the opinion score. Default: \code{1}. See the documentation
of \code{metric} parameter of \code{opi_score} function for details.}

\item{fun}{A user-defined function if parameter \code{metric}
is set as \code{5}. Also, see the documentation
of \code{fun} parameter of \code{opi_score} function for details.}

\item{nsim}{\link{integer} Number of replicas of the OSD to generate.
Recommended values: 99, 999, 9999, and so on. Since the run time
is proportional to the number of replicas, a lower number of
simulation is recommended. Default: \code{99}.}

\item{alternative}{\link{character} By default, this function will
assume a two-tailed test, as indicated by \code{"two.sided"} argument
in the \code{alternative} parameter. You can override
this by specifying \code{“less”} or \code{“greater”}, which will
run the analysis as a one-tailed test with the
criterion (i.e. the observed score) being located at the
lower or upper regions of the distribution, respectively.
Note: when \code{metric} parameter is set as \code{1},
\code{alternative} parameter is \code{"two.sided"}, while it is
\code{"less"} when \code{metric} parameter is set as \code{2},
\code{3}, or \code{4}. For a user-defined metric, i.e.
when \code{metric=5} with \code{fun} parameter specified,
the user needs to determine the bounds of the metric score
(see full documentation of \code{fun} parameter in
\code{opi_score} function). A simple rule that can be
applied is that if the opinion score can assume either
a negative or a positive value, then the \code{alternative}
argument should be set as "two.sided", else if it can
only assume a negative value, it should be set as "less",
and lastly, if it can only assume a positive value, then
it should be set as "greater". Note that the function is
able to detect a contradicting sign to the \code{less}
or \code{greater} argument, and automatically utilize the
\code{"two.sided"} criterion instead.}

\item{pplot}{\link{logical} To display graphical plot showing
the proportion of text records containing (or not
containing) secondary keywords (i.e. 'present' and
'absent' groups), as well as the proportion of
'positive' and 'negative' classes.}
}
\value{
Details of statistical significance of impacts
of subject B on subject A.
}
\description{
To determine the impacts of the opinion
(inherent within a text document) concerning subject B
on the original subject A of the text document. For
Twitter data, subject A refers to the subject whose
related keywords (#hashtags) are used to download the
data. Subject B refers to any other secondary subject
that was being discussed in relation to subject A in the
text document. Keywords relating to secondary subject
need to be identified and input into \code{opi_mpact}
function. These keywords can be identified by running
a preliminary topic analysis on the downloaded text
document. We provided \code{tf_idf} function to help a
user identify important keywords that may signify a
secondary subject in a text document. A user is free to
deploy any other means in order to extract these important
(secondary) keywords. Alternatively, a user can specify
these keywords manually.
}
\details{
This function compares the observed opinion scores
(computed using the \code{opi_score} function) with the
expected opinion scores (distribution) from the
\code{opi_sim} function, in order to estimate the
statistical significance value (p-value). The p-value
is computed as p = (S.beat+1)/(S.total+1),
where 'S_total' is the total number of replicas created,
'S.beat' is number of replicas with the expected score
greater than the observed score. If a user-defined opinion
score function is specified, he/she needs to determine
whether a one-tailed or two-tailed comparison is required.
Typically, a \code{99} replicas is sufficient.
Thus, if, for example, three of the 99
replicas have higher scores than observed score,
then the p-value is equal to (3+1)/(99+1) = 0.04.
}
\references{
(1) Adepeju and Jimoh, 2021,
}
