% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi_sim.R
\name{opi_sim}
\alias{opi_sim}
\title{To simulate the expected sentiment (opinion) distribution}
\usage{
opi_sim(osd_data, nsim=99, metric = 1, fun = NULL, quiet=TRUE)
}
\arguments{
\item{osd_data}{\link{list} An observed sentiment document
of (n x 3) dimension, in which \code{n} is the number of
text records that are successfully classified as
positive, negative or neutral. An \code{osd_data} must be
derived from the text document. Column 1 shows the ID of the
text record, column 2 shows the shows the sentiment
classes of records (i.e. positive, negative, or neutral).
Column 3 indicate the presence  (\code{present}) or absence
(\code{absent}) of secondary keyword in each text record.}

\item{nsim}{\link{integer} Number of replicas of the OSD to generate.
Recommended values: 99, 999, 9999, and so on. Since the run time
is proportional to the number of replicas, a lower number of
simulation is recommended. Default: \code{99}.}

\item{metric}{\link{integer} Metric to utilize for the calculation
of the expected opinion score. Default: \code{1}. See the documentation
of \code{metric} parameter of \code{opi_score} function for details.
To calculate p-values, the argument of \code{metric} parameter
here must correspond to that of \code{opi_score} function.}

\item{fun}{A user-defined function if parameter \code{metric}
is set as \code{5}. Also, see the documentation
of \code{fun} parameter of \code{opi_score} function for details.}

\item{quiet}{(TRUE or FALSE) To suppress processing
messages. Default: \code{TRUE}.}
}
\value{
Returns a list of expected opinion scores based on
the number of simulated, i.e. \code{nsim} specified.
}
\description{
Given a text document with two underlying subjects
A and B, this function simulates (replicates) the opinions associated
with B as the expected opinions (from A). The resulting
text document is referred to as the \verb{Expected Sentiment Document (ESD)} (Adepeju and Jimoh, 2021).
This exercise allow the impacts of B to be tested on A
(see function \code{opi_mpact} for details).
}
\details{
Employs randomization testing approach in
order to generate expected distribution of the opinion scores
that signifies the absence of effects of secondary subject.
Simply, the randomization testing involves generating a large
number of ESD, referred here to as “replicas”, \code{S}.
A replica is generated by randomizing the sentiment assignment of
text document that contain the secondary keywords in accordance
with the probabilities derived from text document that do not
contain those keywords.
}
\references{
(1) Adepeju, M. and Jimoh, F. (2021). An Analytical
Framework for Measuring Inequality in the Public Opinions on
Policing – Assessing the impacts of COVID-19 Pandemic using
Twitter Data. https://doi.org/10.31235/osf.io/c32qh
}
