% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osd_data.R
\docType{data}
\name{osd_data}
\alias{osd_data}
\title{Observed sentiment document (OSD)}
\format{
A list (dataframe)
}
\usage{
osd_data
}
\description{
A tidy-format list (dataframe) showing the
classification of text records into positive, negative
or neutral sentiments. The second column of the dataframe consists of
label variables: \code{present} and \code{absent} which indicate whether any
of the secondary keywords exist in a text record.
}
\keyword{datasets}
