% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policing_otd.R
\docType{data}
\name{policing_otd}
\alias{policing_otd}
\title{Fake Twitter posts on police/policing 1}
\format{
A list (dataframe)
}
\usage{
policing_otd
}
\description{
A sample text document (an OTD) containing twitter posts
(from an anonymous geographical location 1). The primary subject
of the document is the "process of policing". The \code{OTD} is believed
to also include posts that express sentiments on COVID-19 pandemic
(secondary subject B) in relation to the primary subject.
}
\keyword{datasets}
