% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refreshment_theme}
\alias{refreshment_theme}
\title{Keywords relating to facilities at train stations}
\format{
A dataframe containing one variable:
\itemize{
\item keys: list of keywords
}
}
\usage{
refreshment_theme
}
\description{
List of words relating to refreshments that can
be found at the Piccadilly Train
Station (Manchester)
}
\keyword{datasets}
