% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reviews_otd}
\alias{reviews_otd}
\title{Customer reviews from tripadvisor website}
\format{
A dataframe containing one variable
\itemize{
\item text: individual text records
}
}
\usage{
reviews_otd
}
\description{
A text document (an OTD) containing the  customer reviews
of the Piccadilly train station (Manchester) downloaded
from the www.tripadvisor.co.uk'. The reviews cover from
July 2016 to March 2021.
}
\keyword{datasets}
