% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signage_theme}
\alias{signage_theme}
\title{Keywords relating to signages at train stations}
\format{
A dataframe containing one variable:
\itemize{
\item keys: list of keywords
}
}
\usage{
signage_theme
}
\description{
List of signages at the Piccadilly Train
Station (Manchester)
}
\keyword{datasets}
