% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slp.R
\name{testit}
\alias{testit}
\title{yes}
\usage{
testit(x=2)
}
\arguments{
\item{x}{\link{integer}: value ihj jyg}
}
\value{

}
\description{
this
}
\details{
this
}
\examples{
testit(x=2)
}
\references{
kml is one
}
