% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tf_idf}
\alias{tf_idf}
\title{Detect important topics in a text document}
\usage{
tf_idf(textdoc = tweets, grouping=FALSE, n=10, showplot=FALSE)
}
\arguments{
\item{textdoc}{A collection (dataframe) of individual text
records, such as tweets or Facebook posts.
By default: The \code{textdoc} is a one-column dataframe.
If contains two-columns, the second column must contain
text group labels.}

\item{grouping}{Indicate whether the dataframe
includes a second column indicating a defined groupings.
If \code{FALSE}, the function impose arbitrary groups.
Default: \code{FALSE}}

\item{n}{\link{integer} the number of top important words across
the text document to display. Default value: \code{10}}

\item{showplot}{To show graphical plot showing
rank of important words by group. Default: \code{FALSE}}
}
\value{
List of topics with their respective tf-idf values
(importance).
}
\description{
Calculates the tf-idf metric
(Silge & Robinson, 2016) in order to determine
the important topics in a document. The \code{tf-idf}
value represents level of importance of a given
topic (or word) across the group of text that exist in a document.
}
\details{
The idea of tf-idf is to find important
words from the content of the text document across various
text groups by decreasing the weight for commonly used words
and increasing the weight for words that are not used very
much in the documents. Based on the result, a user may be
able to identify words that indicates certain underlying
subjects that relate to the original subject of the text
document. A user can collate these words and use them
as input into the \code{sec_keywords} parameter of the
\code{opi_impact} function. For this function, we built on
the \code{bind_tf_idf} function from the tidytext R-package
(Silge & Robinson, 2016).
}
\references{
Silge, J. and Robinson, D. (2016) tidytext:
Text mining and analysis using tidy data principles in R.
Journal of Open Source Software, 1, 37.
}
