% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets.R
\docType{data}
\name{tweets}
\alias{tweets}
\title{Fake Twitter posts on police/policing 2}
\format{
A list(dataframe)
}
\usage{
tweets
}
\description{
A sample text document (an OTD) containing
twitter posts (from an anonymous geographical location 2).
The primary subject of the document is the “process of
policing” (as the primary subject). The OTD is believed
to also include posts that express sentiments on the
COVID-19 pandemic (Secondary subject B) in relation
to the primary subject.
}
\keyword{datasets}
