% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_keys.R
\docType{data}
\name{covid_keys}
\alias{covid_keys}
\title{COVID-19 pandemic related keywords}
\format{
A list (dataframe)
}
\usage{
covid_keys
}
\description{
A list of keywords relating to the COVID-19 pandemic
}
\keyword{datasets}
