% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{debate_dtd}
\alias{debate_dtd}
\title{Comments on a political debate video}
\format{
A dataframe containing one variable
\itemize{
\item text: individual text records
}
}
\usage{
debate_dtd
}
\description{
A DTD containing individual comments on the video
showing the debate between two US presidential
nominees (Donald Trump and Hillary Clinton)
in Sept. 2016. (Credit: NBC News).
}
\keyword{datasets}
