% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi_impact.R
\name{opi_impact}
\alias{opi_impact}
\title{Impact analysis of subject B on the opinion expressed concerning subject A}
\usage{
opi_impact(textdoc, sec_keywords=NULL, metric = 1,
fun = NULL, nsim = 99, alternative="two.sided",
quiet=TRUE)
}
\arguments{
\item{textdoc}{An \code{n} x \code{1} list (dataframe) of
individual text records, where \code{n} is the total
number of individual records.}

\item{sec_keywords}{\link{list} A one-column dataframe (of any
number of rows) containing a list of keywords relating to
the secondary subject (\code{B}).}

\item{metric}{\link{integer} Metric to utilize for the calculation
of the opinion score. Default: \code{1}. See detailed documentation
in the \code{opi_score} function.}

\item{fun}{A user-defined function provided parameter
\code{metric} is set as \code{5}. See detailed documentation
in the \code{opi_score} function.}

\item{nsim}{\link{integer} Number of replicas (ESD) to generate.
See detailed documentation in the \code{opi_sim} function.
Default: \code{99}.}

\item{alternative}{\link{character} Default: \code{"two.sided"},
indicating a two-tailed test. A user can override
this by specifying \code{“less”} or \code{“greater”} to run
the analysis as a one-tailed test with the observed score
being located at the lower or upper regions of the
distribution, respectively. Note: for \code{metric=1}
(see above), the \code{alternative} parameter should be
set as \code{"two.sided"} because the opinion score is
bounded by both negative and positive values. For a positively
bounded opinion score, such as when
\code{metric = 2, 3 or 4}, the the \code{alternative} parameter
should be "greater", and "less" otherwise.}

\item{quiet}{(TRUE or FALSE) To suppress processing
and warning messages. Default: \code{TRUE}.}
}
\value{
Details of statistical significance of impacts
of secondary subject B on the opinion concerning subject A.
}
\description{
This function assesses the impacts of a
subject B (henceforth referred to as
secondary subject) on the opinion concerning subject A in
a text document. Keywords relating to the secondary subject,
can be identified through a frequency analysis or
collated manually, and should be provided
as input into the function (see below). The subject A
(primary subject) is the subject matter upon which the
text document is based. For instance, by downloading Twitter
data that include hashtags: '#police', '#policing' and/or
'#law enforcement', then "Police/Policing"
becomes the primary subject of the text document.
}
\details{
This function calculate the statistical
significance value (\code{p-value}) by comparing the
observed opinion scores (from the \code{opi_score}
function) with the expected opinion scores (distribution) from the
\code{opi_sim} function. The formula is given as
p = (S.beat+1)/(S.total+1), where 'S_total' is the total
number of replicas created, 'S.beat' is number of replicas
with the expected score greater than the observed score. If a user-defined opinion
score function is specified, he/she needs to determine
whether a one-tailed or two-tailed comparison is required.
(see explanations above).
}
\examples{
#test document: 'policing_otd'
#list of keywords: 'covid_keys'

output <- opi_impact(textdoc = policing_otd,
          sec_keywords=covid_keys, metric = 1,
          fun = NULL, nsim = 99, alternative="two.sided",
          quiet=TRUE)

#check output variables
print(output)

#to access the pvalue
output$pvalue

}
\references{
(1) Adepeju, M. and Jimoh, F. (2021). An Analytical
Framework for Measuring Inequality in the Public Opinions on
Policing – Assessing the impacts of COVID-19 Pandemic using
Twitter Data. https://doi.org/10.31235/osf.io/c32qh
}
