% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi_score.R
\name{opi_score}
\alias{opi_score}
\title{Opinion score of the subject matter in a text document}
\usage{
opi_score(textdoc, metric = 1, fun = NULL)
}
\arguments{
\item{textdoc}{An \code{n} x \code{1} list (dataframe) of
individual text records, where \code{n} is the total
number of individual records.}

\item{metric}{\link{integer} Metric to utilize for the calculation
of the opinion score. Available values are: \code{1, 2, ...,5}.
Assuming \code{P}, \code{N} and \code{O} represent positive,
negative, and neutral text records, respectively, the followings
are few examples of opinion scores from the literature:
\code{1}: Polarity (percentage difference)
\code{((P - N)/(P + N))*100}, (Bound: -100\%, +100\%);
\code{2}: Polarity (proportional difference)
\code{((abs(P - N) / (P + N + O))*100},
(Bound: 0, +100\%);
\code{3}: Positivity \code{(P/ (P + N + O))*100},
(Bound: 0, +100\%); \code{4}: Negativity \code{(N / (P + N + O))*100},
(Bound: 0, +100\%) (Malshe, A. 2019;
Lowe et al. 2011).\code{5}: To pass a
user-defined function as argument into the \code{fun} parameter below.}

\item{fun}{A user-defined function provided parameter
\code{metric} above is set as \code{5}.
For example, given the function \code{myfun} <- function(P, N, O){
("some tasks to do"); return("a value")}, the
\code{fun} parameter is then set as \code{fun = myfun}.
Default: \code{NULL} i.e. when \code{metric} parameter
is not \code{5}.}
}
\value{
Returns an \code{opi_object} containing details of the
opinion measures from the text document.
}
\description{
Given a text document (concerning a subject A),
this function compute the overall opinion score based on the
proportion of text records classified as expressing positive,
negative or neutral sentiment. The function first transforms
the text document into a tidy-format dataframe, referred to
as \verb{observed sentiment document (OSD)}, in which each text
record is assigned a sentiment class based on the sum of all
sentiments expressed by words in the text record.
}
\details{
An opinion score is derived from all the sentiments
(i.e. positive, negative (and neutral) expressed within a
text document. We deploy a lexicon-based approach
(Taboada et al. 2011) using the \code{AFINN} lexicon
(Nielsen, 2011).
}
\references{
(1) Malshe, A. (2019) Data Analytics Applications.
Online book available at:
https://ashgreat.github.io/analyticsAppBook/index.html.
Date accessed: 15th December 2020.
(2) Taboada, M.et al. (2011).
Lexicon-based methods for sentiment analysis. Computational
linguistics, 37(2), pp.267-307.
(3) Lowe, W. et al. (2011).
Scaling policy preferences from coded political texts.
Legislative studies quarterly, 36(1), pp.123-155.
}
