% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osd_data.R
\docType{data}
\name{osd_data}
\alias{osd_data}
\title{Observed sentiment document (OSD)}
\format{
A list (dataframe)
}
\usage{
osd_data
}
\description{
A tidy-format list (dataframe) showing the resulting
classification of each text records into positive, negative
or neutral sentiment. The second column of the dataframe consists of
labels variables \code{present} and \code{absent} to indicate whether any
of the secondary keywords exist in a text record.
}
\keyword{datasets}
