% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osd_data.R
\docType{data}
\name{osd_data}
\alias{osd_data}
\title{observed sentiment document}
\format{
A list
}
\usage{
osd_data
}
\description{
A tidy-format list showing the resulting
classification of eligible text records as positive, negative
or neutral, with their respective \code{present} or \code{absent}
labels to indicate the presence or absence of any secondary
keywords in a text record.
}
\keyword{datasets}
