% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{osd_data}
\alias{osd_data}
\title{Observed sentiment document (OSD).}
\format{
A dataframe with the following variables:
\itemize{
\item ID: numeric id of text record with valid
resultant sentiments score and classification.
\item sentiment: Containing the sentiment classes.
\item keywords: Indicator to show whether a secondary
keyword in present or absent in a text record.
}
}
\usage{
osd_data
}
\description{
A tidy-format list (dataframe) showing the resulting
classification of each text record into positive, negative
or neutral sentiment. The second column of the dataframe consists of
labels variables \code{present} and \code{absent} to indicate whether any of the secondary
keywords exist in a text record.
}
\keyword{datasets}
