% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policing_dtd}
\alias{policing_dtd}
\title{Fake Twitter posts on police/policing}
\format{
A dataframe containing one variable
\itemize{
\item text: individual text records
}
}
\usage{
policing_dtd
}
\description{
A text document (an OTD) containing twitter posts
(for an anonymous geographical location 1) on police/policing.
The OTD also includes
posts that express sentiments on policing in relation to
the COVID-19 pandemic (Secondary subject B)
}
\keyword{datasets}
