% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policing_otd.R
\docType{data}
\name{policing_otd}
\alias{policing_otd}
\title{Sample simulated Twitter posts}
\format{
A matrix
}
\usage{
policing_otd
}
\description{
This text document contains simulated twitter posts
concerning policing. Within this OTD, a number of posts express
certain sentiments COVID-19 pandemic in relation to police/policing
activities within an area.
}
\keyword{datasets}
