% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policing_otd.R
\docType{data}
\name{policing_otd}
\alias{policing_otd}
\title{Fake Twitter posts on police/policing 1}
\format{
A list (dataframe)
}
\usage{
policing_otd
}
\description{
A text document (an OTD) containing twitter posts
(for an anonymous geographical location 1) on police/policing
(primary subject A). The OTD includes
posts that express sentiments on policing in relation to
the COVID-19 pandemic (Secondary subject B)
}
\keyword{datasets}
