% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf_idf.R
\name{tf_idf}
\alias{tf_idf}
\title{Highlight the top-most important topics in a text document}
\usage{
tf_idf(textdoc, n_top=10, showplot=FALSE)
}
\arguments{
\item{textdoc}{An \code{n} x \code{1} list (dataframe) of
individual text records, where \code{n} is the total
number of individual records. An \code{n} x code{2} dataframe can
also be supplied, but the second column must contain
pre-defined group labels of the text records, e.g. group labels
according to geographical locations.
For an \code{n} x code{1} dataframe, the function
automatically impose arbitrary group labels on
the text document, based on the length \code{n} of the document.
Default: \code{FALSE}}

\item{n_top}{\link{integer} maximum number of top-most important words
per group to display. Default value: \code{10}}

\item{showplot}{To display graphical plot showing
ranks of top-most important words by group.
Default: \code{FALSE}}
}
\value{
A graphical display showing top-most important
words in each group, according to the \code{tf-idf}
measure.
}
\description{
This function identifies the most
important words across different text groups in a text document,
according to the \code{tf-idf} measure (Silge & Robinson, 2016).
}
\details{
The function utilizes the \code{tf-idf} measure
in order to determine the most important words across various
text groups in a document. The idea of \code{tf-idf} is
to find words that are not used very much, but appear across
many groups in the document. using this function, a user
may be able to identify certain keywords that indicates some
underlying (secondary) subject that are been discussed
in relation to the original (primary) subject of the text
document.
}
\references{
Silge, J. and Robinson, D. (2016) tidytext:
Text mining and analysis using tidy data principles in R.
Journal of Open Source Software, 1, 37.
}
