% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweets.R
\docType{data}
\name{tweets}
\alias{tweets}
\title{Fake Twitter posts on police/policing 2}
\format{
A list(dataframe)
}
\usage{
tweets
}
\description{
A text document (an OTD) containing twitter posts
(for an anonymous geographical location 2) on police/policing
(primary subject A). The OTD includes
posts that express sentiments on policing in relation to
the COVID-19 pandemic (Secondary subject B)
}
\keyword{datasets}
