% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tweets}
\alias{tweets}
\title{Fake Twitter posts on police/policing 2}
\format{
A dataframe with the following variables:
\itemize{
\item text: individual text records
\item group: real/arbitrary groups of text records
}
}
\usage{
tweets
}
\description{
A text document (an DTD) containing twitter posts
(for an anonymous geographical location 2) on police/policing
(primary subject A). The DTD includes
posts that express sentiments on policing in relation to
the COVID-19 pandemic (Secondary subject B)
}
\keyword{datasets}
