% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/word_distrib.R
\name{word_distrib}
\alias{word_distrib}
\title{Term Distribution}
\usage{
word_distrib(textdoc)
}
\arguments{
\item{textdoc}{\code{n} x \code{1} list (dataframe)
of individual text records, where \code{n} is the number
of individual records.}
}
\value{
A graphical plot showing the rank-frequency graph.
}
\description{
This function examines whether the word
frequency in a text document follows the Zipf distribution
(Zipf 1934). The Zipf's distribution is considered the
distribution of a perfect natural language text.
}
\details{
The Zipf's distribution is most easily observed by
plotting the data on a log-log graph, with the axes being
log(word rank order) and log(word frequency). For a perfect
natural language text, the relationship between the rank and
the frequency should have a negative slope with all points
falling on a straight line. Any deviation from a straight
line can be considered an element of imperfection from the
text document.
}
\examples{

#Get an \code{n} x 1 text document
tweets_dat <- as.data.frame(tweets[,1])
plt = word_distrib(textdoc = tweets_dat)

plt

}
\references{
Zipf G (1936). The Psychobiology of Language.
London: Routledge; 1936.
}
